requirements.txt
pandas==1.4.2
Pillow==9.2.0
Pillow==10.4.0
pymongo==4.8.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.37.0
streamlit_option_menu==0.3.13
